feat(setup_shell): Add custom fish prompt
The custom fish prompt will look like this:

- Non git dir: `[HH:MM] user@host:short_pwd ▶ `
- Git directory: `[HH:MM] user@host:short_pwd (branch) ▶ `

Also added a setp to deploy the static fish configs to the playbook
targeting GameMaster. This should've been a separate commit, but for now
deploying the playbook is the only way I can manually test.

For now this is just fine, what I'm thinking is to design a two line
prompt that also provides more project information, much more than the
git branch, but that's a bridge I'm not ready to cross yet.