feat(test): Add ArchLinux test container file
Adapted the ArchLinux test file like I did for Fedora.

First we have a container file that dictates how the ArchLinux base test
image should be built, containing an installation of Ansible and the
"droptix" user.

Next we have the corresponding shell script that was changed to take
advantage of the ArchLinux container file. It builds the image based on
it and starts an interactive shell in the container with the CWD set to
`/opt/system-forger`.