feat(scripts): Add Ansible installation bootstrap script
Introduces `./scripts/install_ansible.sh`, a helper script to install
Ansible on the target system.

Because we target Fedora, Ubuntu and ArchLinux we check for the
existance of their corresponding package managers (dnf, apt or pacman)
and installs Ansible. It is idempotent only installing Ansible if not
already present.

The script is a convenient way to preapare a new machine for Ansible
playbook execution.