feat(test): Install sudo in Ubuntu 24.04 test container
By default `sudo` is not available in the Ubuntu 24.04 container image.

There are scripts like `./scripts/install_ansible.sh` that are expecting
the existance of `sudo`. Because of this we add it to the container.

Another small change added with this commit is the simplifaction of
logs. The `---` decorations were removed.