feat(test): Remove ansible instalation from all test container scripts
The idea is to bootstrap Ansible using `./scripts/install_ansible.sh`
script. If we do it as a bootstrap step of the test container, then to
test the Ansible bootstrap script we need to first uninstall Ansible.
This adds friction that feels unnatural.

Having to run the Ansible bootstrap script by hand adds friction too,
but I feel it is better.