feat(setup_shell): Define per distribution packages list
There are cases when different distros use different names for the same
packages.

One solution is to have multiple `package` tasks, one for the common
packages, one for particular packages for Ubuntu and so on.

The current solution where we define in the default variables the list
of packages per distribution, I find it more friendly and easy to
maintain.