feat(setup_rust): Install rust using rustup
The installation is done using the official https://rustup.rs installer.
We simply download the bash script to a temporary file and run it.

The installation is done only if `$HOME/.cargo/bin/rustc` file does not
exist. This is the way we add idempotency to the rust installation. Is
it enough I don't know, but as a starting point it is perfect.

By default the installer also changes the `PATH`, but we intentionally
disable that. In a future commit we'll take care of loading the rust
environment.