feat(setup_rust): Add bash and fish environment loaders for Rust
Because we want to control how the environment is setup we prefer to add
`rust_env.sh` in `$HOME/.bashrc.d/` for bash and `rust_env.fish` in
`$HOME/.config/fish/conf.d/` for fish.

Those are modular loaders for Rust environment. Those will be
automatically sourced by the shells and will help by not requiring
changes to the main shell config files `.bashrc` and `config.fish`.