feat(setup_shell): Make sure `.bashrc.d/` and `.config/fish/conf.d` exist
Those locations will be used by other Ansible roles to place shell
modules. One such example when this is done is in `setup_rust` where we
deploy the Rust environment loaders in those locations.

It's the job of `setup_shell` to prepare the locations and for other
roles to assume they exist. It's not worth duplicating the check for the
existance of those locations in both `setup_shell` and any other role
that requires them.