feat(playbooks): Add `workstation` and `wsl` playbooks
Creating `ansible.cfg` to define the `roles_path` so playbooks will be
able to find the roles in the future as I took the decision to store
roles in `./roles` and non-role tasks in `./tasks`, relative to the
project root. Last, but not least, the playbooks will be sored in
`./playbooks`.

Introduce the foundational playbooks, `playbooks/workstation.yml` and
`playbooks/wsl.yml`, which serve as the primary entry points for
configuring respective environments. The big picture is like this, use
`workstation` playbook to setup bare metal installations or VM
installations. The `wsl` playbook is meant more for a Windows Substystem
for Linux, but should be safe to apply to any CLI only deployment
containers, WSL itself or VMs where CLI only setup is desired. I'll
think about using a better name, but for now it stays like this.

Both playbooks are set up to target `localhost` and include a
`pre_tasks` section dedicated to gathering facts about the target user
and validating their existence, providing a consistent user context for
future configuration tasks.