feat(setup_base_cli): Add `fish` installation and configuration
Fish will be used as interactive shell for the user. For interactive
prompts is so much better than `bash`. Shell scripting is still meant to
be done using `bash`, but for interactive work we go the `fish` way.

This includes:

- Installation of the `fish` package across supported distributions.

- Conditional installation of the `hostname` utility based on
  distribution (`inetutils` for ArchLinux, `hostname` for Fedora and
  Ubuntu).

- Ensuring necessary Fish configuration directories for modular
  configuratons  (`~/.config/fish/conf.d/`) and the ones for custom
  functions (`~/.config/fish/functions/`) exist.

- Deployment of a custom `fish_prompt` providing a personalized
  command-line prompt (`[HH:MM] user@host:cwd (git branch) |`).

  The git branch is shown only when inside a git repository and `cwd` is
  the name of the CWD, not its path.

The `roles/setup_base_cli/tasks/main.yml` is updated to include these
new Fish-related setup tasks.