refactor(setup_base_cli): Simplify `fish` shell directory and function deployment
In this simple refactoring two things are happening:

- Merges the creation of `$HOME/.config/fish/conf.d/` and
  `$HOME/.config/fish/functions/` into a single task using a loop.

- Merges the deployment of `fish_prompt.fish` and `fish_greeting.fish`
  into a single copy task using a loop

Using this pattern if we want to deploy another file further down the
line we just add the file in `files` and add another item to the loop.

Another soliton would be to just copy the entire `.config/fish` into
`$HOME/.config/fish`, but here we prefer specificity.