feat(setup_base_cli): Add `eza` alternative to `ls`
For Ubuntu and ArchLinux it is easy, just use the version provided by
the package manager.

For Fedora there is another story. Since Fedora 42+ `eza` package is no
longer maintained. This means no `eza` in `dnf`.

For us the best decision for now is to download the latest `eza` tarball
from their GitHub page and install that. The way we implemented this we
don't take care of idempotency. It will always download tarball,
extract, install `eza` to `/usr/bine/eza`.