feat(scripts): Add 'install.sh' and moved all scripts to './scripts'
The 'install.sh' script is needed because it simplifies the call to
'ansible-playbook'. Handling the call manually is problematic especially
when you have to differentiate between manual runs, test container vs
proper VM or machine etc.

The script hiddes all the complexity related to asking for the become
password (think of it as the 'sudo' password).

Another argument for it is ergonomics. The call to 'ansible-playbook' is
defined in the script and no matter where we need to run it the command
and behavior will be the same: run from 'bootstrap.sh' or run from the
terminal by the user.