fix(test): Mount everything except '.git'
To test 'system-forger' we mount the current git tree in the test
container. If the current tree is a work tree where '.git' is a file,
NOT a directory, and it points to a non-existing path from the point of
view of 'git', then running 'git' commands from inside the mounted
'system-forger' tree will result in errors like this one:

    # git ls-remote https://github.com/DrOptix/system-forger
    fatal: not a git repository:/tmp/git_ls_remote_bug/.bare/worktrees/master

Now with our strategy of mounting everyting except '.git' we work around
the issue, not the nicest, it might even add slight delays for container
initialization.

One more thing. I'm not sure how it will evolve, but I wrote about
this bug, it's a bug in 'git' in my opinion, on the 'git mailing list.

Here it is and lets see how it goes:
https://lore.kernel.org/git/6be0928c-df10-417e-a264-f4562f9d2585@gmail.com/T/#u