fix(setup_base_cli): Correct 'bash' config deployment and refactor tasks
Fix a critical bug in the 'bash' configuration deployment where
'.bash_profile' was being incorrectly deployed with the content of
'.bashrc'.

The 'src' path in the file deployment task is now dynamically
set ('bash/{{ item }}'), ensuring '.bashrc' and '.bash_profile' are
copied from their respective correct source files.

Alongside this fix, the 'bash' tasks are refactored.

Directory creation is consolidated into a single task:

    Ensure required directories exists

The same thing for configuration files, using the task:

    Deploy configuration files

Task names are made more generic to reflect these consolidated
operations. From now on when we need to deploy a new configuration file
or make sure a directory exists for 'bash' we just have to add a new
item to the loop.