feat(setup_base_cli): Add file fuzzy finder
Use 'telescope' plugin to provide an ergonomic fuzzy finder.

To not manager plugins by hand we make use of 'lazy.nvim' plugin
manager.

The plugins are automatically installed when 'neovim' starts for the
first time. Not always the case, but sometimes 'neovim' requires a
restart after the plugins have been installed.