test(bash): Add molecule test
The role of this battery of tests is to make sure the 'bash' role
follows our design phylosophy even after N years.

The way the 'verify' playbook is written kind of makes me puke, but I
couldn't figure out a cleaner way to do it. For now it is good enough
and it will provide a good base for the next 'verify' playbooks. I'll
have to write.